apply the new editing backend is assigned + blocked building + first login + change password

+++ ui/wizards/office_survey/steps/relation_step.py
@@ -445,7 +445,7 @@ class RelationStep(BaseStep):
                 "person_id": person_data.get('person_id'),
                 "person_name": f"{person_data.get('first_name', '')} {person_data.get('last_name', '')}",
                 "relation_type": card.relation_type.currentData() if hasattr(card.relation_type, 'currentData') else None,
-                "ownership_share": float(card.ownership_share.text()) if card.ownership_share.text() and card.ownership_share.text() != "0" else 0.0,
+                "ownership_share": int(card.ownership_share.text()) if card.ownership_share.text() and card.ownership_share.text() != "0" else 0,
                 "start_date": card.start_date.date().toPyDate().isoformat(),
                 "contract_type": card.contract_type.currentData() if card.contract_type.currentIndex() > 0 else None,
                 "evidence_type": card.evidence_type.currentData() if card.evidence_type.currentIndex() > 0 else None,
